Create Symantec EDR
To parse your Symantec EDR logs (JSON), You need to collect the logs, create a new custom logs in Azure Sentinel portal and create a new function to parse the logs
 
Data collection:

From Symantec EDR forward your logs to your log collector through port 514

How to create a custom logs:

Step 1. Open the Custom Log Wizard

The Custom Log Wizard runs in the Azure portal and allows you to define a new custom log to collect.

    In the Azure portal, select Log Analytics workspaces > your workspace > Settings.
    Click on Custom logs.
    By default, all configuration changes are automatically pushed to all agents. For Linux agents, a configuration file is sent to the Fluentd data collector.
    Click Add+ to open the Custom Log Wizard.

Step 2. Upload and parse a sample log

You start by uploading a sample of the custom log. The wizard will parse and display the entries in this file for you to validate. Azure Monitor will use the delimiter that you specify to identify each record.

New Line is the default delimiter and is used for log files that have a single entry per line. If the line starts with a date and time in one of the available formats, then you can specify a Timestamp delimiter which supports entries that span more than one line.

If a timestamp delimiter is used, then the TimeGenerated property of each record stored in Azure Monitor will be populated with the date/time specified for that entry in the log file. If a new line delimiter is used, then TimeGenerated is populated with date and time that Azure Monitor collected the entry.

    Click Browse and browse to a sample file. Note that this may button may be labeled Choose File in some browsers.
    Click Next.
    The Custom Log Wizard will upload the file and list the records that it identifies.
    Change the delimiter that is used to identify a new record and select the delimiter that best identifies the records in your log file.
    Click Next.

Step 3. Add log collection paths

You must define one or more paths on the agent where it can locate the custom log. You can either provide a specific path and name for the log file, or you can specify a path with a wildcard for the name. This supports applications that create a new file each day or when one file reaches a certain size. You can also provide multiple paths for a single log file.

For example, an application might create a log file each day with the date included in the name as in log20100316.txt. A pattern for such a log might be log*.txt which would apply to any log file following the application’s naming scheme.

The following table provides examples of valid patterns to specify different log files.

Description 	Path
All files in C:\Logs with .txt extension on Windows agent 	C:\Logs\*.txt
All files in C:\Logs with a name starting with log and a .txt extension on Windows agent 	C:\Logs\log*.txt
All files in /var/log/audit with .txt extension on Linux agent 	/var/log/audit/*.txt
All files in /var/log/audit with a name starting with log and a .txt extension on Linux agent 	/var/log/audit/log*.txt
(All files in /var/log/symantec-edr* with a name starting with symantec-edr on Linux log collector) 

    Select Windows or Linux to specify which path format you are adding.
    Type in the path and click the + button.
    Repeat the process for any additional paths.

Step 4. Provide a name and description for the log "symantec_edr"
 
The name that you specify will be used for the log type as described above. It will always end with _CL to distinguish it as a custom log.

    Type in a name for the log. The _CL suffix is automatically provided.
    Add an optional Description.
    Click Next to save the custom log definition.

Step 5. Validate that the custom logs are being collected

It may take up to an hour for the initial data from a new custom log to appear in Azure Monitor. It will start collecting entries from the logs found in the path you specified from the point that you defined the custom log. It will not retain the entries that you uploaded during the custom log creation, but it will collect already existing entries in the log files that it locates.

Once Azure Monitor starts collecting from the custom log, its records will be available with a log query. Use the name that you gave the custom log as the Type in your query.

Note

If the RawData property is missing from the query, you may need to close and reopen your browser.

*Reference: https://docs.microsoft.com/en-us/azure/azure-monitor/agents/data-sources-custom-logs

Then you need to create a new function:

1-Go to the sentinel workspace -> Logs 
2-Write the mentioned query
3-Save -> save as a function.

Thank You.